move some configs
